Move login_api into users app
Now everything is related to authentication is contained within the
users app.
